(REFACTOR) user/mypage 구조 통일, 컴포넌트 분리 및 Contract 생성 분리 (#20)
